Server Rendering Complete
I got react rendering on the server and then getting taken over on the client.
I have worked in passing the state from redux from the server to the client. At this point I have to make everything immutable once it gets to the client side. Seems wrong to do it this manual way.