Server rendering done
I have react rendering on the server and passing the redux store along. There is an example of an action dispatch before the app loads on the client. This is to simulate a login type of event. I have a 'protected' route but I don't think it is the right way given the API on react-router, specifically `onEnter` property of the Route component.

I will end this package here and fork to add in react-redux-router to better handle routing on the client side. Then I will try and get redux-auth-wrapper working with passport and mongoose to deal with login and logout.

The goal is to dispatch an action when the form submits which will POST to the server. The reducer on the client side will be set up to handle the result from the server. Therefore the res.json() object will be an action object with a type and payload properties.